requirements.txt
tensorflow
streamlit
pillow
numpy
pandas
keras